+++ src/apis/client/deleteCard.ts
@@ -1,4 +1,4 @@
-import { cardFetcher } from '../utils/cardFetcher';
+import { cardFetcher } from '../fetcher/cardFetcher';
 
 export const deleteCard = async (id: string) => {
     return await cardFetcher(`/cards/${id}`, { method: 'DELETE' });
